requirements.txt
Django==1.9
django-crispy-forms==1.5.2
django-filter==0.11.0
django-tables2==1.0.5
mysql-connector-python==2.0.4
mysql-connector-python-rf==2.1.3
requests==2.9.0
six==1.10.0
wheel==0.24.0
